Moved iomediator00 into iomediator package.
